FIN-API-4: move service handling files and models to their respective directories
