+++ main.py
@@ -1,9 +1,9 @@
-from typing import Any, Dict, List, Union, Optional
-from fastapi import FastAPI, HTTPException, Query
+from typing import Dict, List, Union
+from fastapi import FastAPI, Query
 from fastapi.middleware.cors import CORSMiddleware
-import requests
-from pydantic import BaseModel
-from config import API_KEY, FINANCIAL_API_BASE_URL
+from models.historical_price import HistoricalPriceResponse, TickerData
+from services.historical_price_service import get_historical_price
+from services.ticker_service import get_ticker_list
 
 app = FastAPI()
 
@@ -15,46 +15,14 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
-FINANCIAL_API_HISTORICAL_URL = f"{FINANCIAL_API_BASE_URL}/historical-price-full"
-SEARCH_TICKER_URL = f"{FINANCIAL_API_BASE_URL}/search-ticker"
-
 
 @app.get("/")
 async def read_root() -> Dict[str, str]:
     return {"Health": "OK"}
 
 
-class HistoricalPrice(BaseModel):
-    date: str
-    open: float
-    high: float
-    low: float
-    close: float
-    adjClose: float
-    change: float
-    changeOverTime: float
-    changePercent: float
-    unadjustedVolume: int
-    volume: int
-    vwap: float
-    label: str
-
-
-class HistoricalPriceResponse(BaseModel):
-    symbol: Optional[str]
-    historical: Optional[List[HistoricalPrice]]
-
-
-class TickerData(BaseModel):
-    symbol: str
-    name: str
-    currency: Optional[str]
-    stockExchange: str
-    exchangeShortName: str
-
-
 @app.get("/historical-price", response_model=HistoricalPriceResponse | Dict[None, None])
-def get_historical_price(
+def handle_historical_price_request(
     symbol: str = Query(
         ..., title="Symbol", description="Symbol of the financial instrument."
     ),
@@ -65,44 +33,11 @@ def get_historical_price(
         ..., title="End Date", description="End date for historical prices."
     ),
 ) -> Union[HistoricalPriceResponse, Dict[None, None]]:
-    params = {
-        "apikey": API_KEY,
-        "from": date_start,
-        "to": date_end,
-    }
-
-    response = requests.get(
-        f"{FINANCIAL_API_HISTORICAL_URL}/{symbol}",
-        params=params,
-    )
-
-    if response.status_code == 200:
-        data = response.json()
-        if "historical" in data and data["historical"]:
-            return HistoricalPriceResponse(symbol=symbol, historical=data["historical"])
-        else:
-            return {}
-    elif response.status_code == 401:
-        raise HTTPException(status_code=401, detail="Invalid API key")
-    elif response.status_code == 404:
-        return {}
-    else:
-        raise HTTPException(status_code=response.status_code, detail="Unexpected error")
+    return get_historical_price(symbol, date_start, date_end)
 
 
 @app.get("/search-ticker", response_model=List[TickerData])
-def get_ticker_list(
+def handle_ticker_search_request(
     query: str = Query(..., title="Query", description="Query for the ticker list.")
 ) -> List[TickerData]:
-    params = {
-        "apikey": API_KEY,
-        "query": query,
-    }
-
-    response = requests.get(SEARCH_TICKER_URL, params=params)
-
-    if response.status_code == 200:
-        data: List[TickerData] = response.json()
-        return data
-    else:
-        raise HTTPException(status_code=response.status_code, detail="Unexpected error")
+    return get_ticker_list(query)
