+++ models/ticker_data.py
@@ -1,0 +1,10 @@
+from typing import Optional
+from pydantic import BaseModel
+
+
+class TickerData(BaseModel):
+    symbol: str
+    name: str
+    currency: Optional[str]
+    stockExchange: str
+    exchangeShortName: str
